Patika Update And Delete Features is Added

+++ com/clonepatika/View/OperatorGUI.java
@@ -9,8 +9,8 @@ import com.clonepatika.Model.User;
 import javax.swing.*;
 import javax.swing.event.TableModelEvent;
 import javax.swing.table.DefaultTableModel;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
+import java.awt.*;
+import java.awt.event.*;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
@@ -40,11 +40,15 @@ public class OperatorGUI extends JFrame {
     private JPanel pnl_adduser;
     private JTable tbl_patikalist;
     private JScrollPane scrll_patikalist;
+    private JPanel pnl_addpatika;
+    private JTextField fld_patikaname;
+    private JButton btn_addpatika;
     private DefaultTableModel mdl_userlist;
     private Object[] row_userlist;
     private final Operator operator;
     private DefaultTableModel mdl_patikalist;
     private Object[] row_patikalist;
+    private JPopupMenu patikaMenu;
 
     public OperatorGUI(Operator operator) throws SQLException {
         this.operator = operator;
@@ -113,19 +117,75 @@ public class OperatorGUI extends JFrame {
 
             }
         });
+        // ## ModelUserList
 
-        mdl_patikalist=new DefaultTableModel(){
+        // ModelPatikaList
+        patikaMenu = new JPopupMenu();
+        JMenuItem updateMenu = new JMenuItem("Güncelle");
+        JMenuItem deleteMenu = new JMenuItem("Sil");
+        patikaMenu.add(updateMenu);
+        patikaMenu.add(deleteMenu);
+        tbl_patikalist.setComponentPopupMenu(patikaMenu);
+
+        updateMenu.addActionListener(e -> {
+            int selectedId = Integer.parseInt(tbl_patikalist.getValueAt(tbl_patikalist.getSelectedRow(), 0).toString());
+
+            try {
+                UpdatePatikaGUI updatePatikaGUI = new UpdatePatikaGUI(Patika.getFetch(selectedId));
+                updatePatikaGUI.addWindowListener(new WindowAdapter() {
+                    @Override
+                    public void windowClosed(WindowEvent e) {
+                        try {
+                            loadPatikaModel();
+                        } catch (SQLException ex) {
+                            ex.printStackTrace();
+                        }
+                    }
+                });
+            } catch (SQLException ex) {
+                ex.printStackTrace();
+            }
+        });
+        deleteMenu.addActionListener(e -> {
+            int selectedId = Integer.parseInt(tbl_patikalist.getValueAt(tbl_patikalist.getSelectedRow(), 0).toString());
+            if (Helper.confirm("sure")) {
+                try {
+                    if (Patika.deletePatika(selectedId)) {
+                        Helper.showMessage("done");
+                        loadPatikaModel();
+                    } else {
+                        Helper.showMessage("error");
+                    }
+                } catch (SQLException ex) {
+                    ex.printStackTrace();
+                }
+            }
+        });
+
+        mdl_patikalist = new DefaultTableModel() {
             @Override
             public boolean isCellEditable(int row, int column) {
                 if (column == 0) return false;
                 return super.isCellEditable(row, column);
             }
         };
-        Object[] col_patikalist={"ID","Patika Adı"};
+
+        Object[] col_patikalist = {"ID", "Patika Adı"};
         mdl_patikalist.setColumnIdentifiers(col_patikalist);
         tbl_patikalist.setModel(mdl_patikalist);
         tbl_patikalist.getTableHeader().setReorderingAllowed(false);
+        tbl_patikalist.getColumnModel().getColumn(0).setMaxWidth(75);
+        tbl_patikalist.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mousePressed(MouseEvent e) {
+                Point point = e.getPoint();
+                int selectedRow = tbl_patikalist.rowAtPoint(point);
+                tbl_patikalist.setRowSelectionInterval(selectedRow, selectedRow);
+
+            }
+        });
         loadPatikaModel();
+        // ## ModelPatikaList
 
         btn_adduser.addActionListener((e) -> {
             if (Helper.isFieldEmpty(fld_name) || Helper.isFieldEmpty(fld_username) || Helper.isFieldEmpty(fld_password)) {
@@ -146,17 +206,18 @@ public class OperatorGUI extends JFrame {
             }
         });
         btn_deleteuser.addActionListener(e -> {
-
-            try {
-                if (User.deleteUser(Integer.parseInt(fld_deleteuser.getText()))) {
-                    Helper.showMessage("done");
-                    loadUserModel();
-                    Helper.clearFields(fld_deleteuser);
-                } else {
-                    Helper.showMessage("error");
+            if (Helper.confirm("sure")) {
+                try {
+                    if (User.deleteUser(Integer.parseInt(fld_deleteuser.getText()))) {
+                        Helper.showMessage("done");
+                        loadUserModel();
+                        Helper.clearFields(fld_deleteuser);
+                    } else {
+                        Helper.showMessage("error");
+                    }
+                } catch (SQLException ex) {
+                    ex.printStackTrace();
                 }
-            } catch (SQLException ex) {
-                ex.printStackTrace();
             }
         });
 
@@ -165,7 +226,7 @@ public class OperatorGUI extends JFrame {
             String username = fld_username_search.getText();
             String type = cbx_usertype_search.getSelectedItem().toString();
             try {
-                ArrayList<User> listUsers=User.searchUser(User.searchQuery(name, username, type));
+                ArrayList<User> listUsers = User.searchUser(User.searchQuery(name, username, type));
                 loadUserModel(listUsers);
 
             } catch (SQLException ex) {
@@ -177,6 +238,19 @@ public class OperatorGUI extends JFrame {
             @Override
             public void actionPerformed(ActionEvent e) {
                 dispose();
+            }
+        });
+        btn_addpatika.addActionListener(e -> {
+            Patika patika = new Patika(0, fld_patikaname.getText());
+            try {
+                if (Patika.addPatika(patika)) {
+                    Helper.showMessage("done");
+                    loadPatikaModel();
+                } else {
+                    Helper.showMessage("error");
+                }
+            } catch (SQLException ex) {
+                ex.printStackTrace();
             }
         });
     }
@@ -198,11 +272,12 @@ public class OperatorGUI extends JFrame {
             mdl_userlist.addRow(row_userlist);
         }
     }
-    public void loadPatikaModel() throws SQLException{
-        DefaultTableModel clearModel= (DefaultTableModel) tbl_patikalist.getModel();
+
+    public void loadPatikaModel() throws SQLException {
+        DefaultTableModel clearModel = (DefaultTableModel) tbl_patikalist.getModel();
         clearModel.setRowCount(0);
         for (Patika patika : Patika.showPatikaList()) {
-            row_patikalist = new Object[]{patika.getId(),patika.getName()};
+            row_patikalist = new Object[]{patika.getId(), patika.getName()};
             mdl_patikalist.addRow(row_patikalist);
         }
